+++ js/application.js
@@ -25,22 +25,29 @@
         {
             case 'cat/lod':
                 {
-                    // parse the level
-                    var level = jsJOMlib__getParameterByName(settings.url, 'l');
-                    if ( parseInt(level, 10) === 1 ) {
-                        JOM['new_job'].categories.categories      = JSON_response.data;
-                        JOM['new_job'].set_categories_list();
-                        JOM['new_job'].categories.nonce.nonce     = JSON_response.new_nonce;
-                        JOM['new_job'].categories.nonce.timestamp = JSON_response.new_timestamp;
+                    if ( JSON_response.c == 'new_job' )
+                    {
+                        // parse the level
+                        var level = jsJOMlib__getParameterByName(settings.url, 'l');
+                        if ( parseInt(level, 10) === 1 ) {
+                            JOM['new_job'].categories.categories      = JSON_response.data;
+                            JOM['new_job'].set_categories_list();
+                            JOM['new_job'].categories.nonce.nonce     = JSON_response.new_nonce;
+                            JOM['new_job'].categories.nonce.timestamp = JSON_response.new_timestamp;
+                        }
+                        else
+                        if ( parseInt(level, 10) === 2 ) {
+                            JOM['new_job'].issues.categories      = JSON_response.data;
+                            JOM['new_job'].set_issues_list();
+                            JOM['new_job'].issues.nonce.nonce     = JSON_response.new_nonce;
+                            JOM['new_job'].issues.nonce.timestamp = JSON_response.new_timestamp;
+                        }
                     }
                     else
-                    if ( parseInt(level, 10) === 2 ) {
-                        JOM['new_job'].issues.categories      = JSON_response.data;
-                        JOM['new_job'].set_issues_list();
-                        JOM['new_job'].issues.nonce.nonce     = JSON_response.new_nonce;
-                        JOM['new_job'].issues.nonce.timestamp = JSON_response.new_timestamp;
-                    }
+                    if ( JSON_response.c == 'search_filter' )
+                    {
 
+                    }
                     break;
                 }
             case 'job/new':
@@ -64,6 +71,8 @@
             case 'sta/lod':
                 {
                     JOM.search_filters.filters['filter by status'].GUI__set_statuses_data(JSON_response.data);
+                    JOM.search_filters.filters['filter by status'].nonce.nonce      = JSON_response.new_nonce;
+                    JOM.search_filters.filters['filter by status'].nonce.timestap   = JSON_response.new_timestamp;
                     break;
                 }